:art: Rename utils2 to utils
