+++ app/src/main/java/com/kyhsgeekcode/disassembler/utils/Parameter.java
@@ -1,0 +1,4 @@
+package com.kyhsgeekcode.disassembler.utils;
+
+class Parameter {
+}
