+++ app/src/main/java/com/kyhsgeekcode/disassembler/utils/Prototype.java
@@ -1,4 +1,4 @@
-package com.kyhsgeekcode.disassembler.utils2;
+package com.kyhsgeekcode.disassembler.utils;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -8,7 +8,7 @@ public class Prototype {
     public String returnType;
     public List<Parameter> params = new ArrayList<>();
 
-    public class Parameter extends com.kyhsgeekcode.disassembler.utils2.Parameter {
+    public class Parameter extends com.kyhsgeekcode.disassembler.utils.Parameter {
         String type;
         String name;
     }
