--- app/src/main/java/com/kyhsgeekcode/disassembler/utils2/Parameter.java
@@ -1,4 +1,0 @@
-package com.kyhsgeekcode.disassembler.utils2;
-
-class Parameter {
-}
