--- app/src/main/java/com/kyhsgeekcode/disassembler/utils2/headerparser/Parser.java
@@ -1,5 +1,0 @@
-package com.kyhsgeekcode.disassembler.utils2.headerparser;
-
-public class Parser {
-
-}
